Move lexer into separate module. Convert crate into library project.
